refactor(docs): 统一 LLM 文档生成插件至 common/docs
