+++ packages/tdesign-uniapp-chat/site/web/plugins/generate-llms/index.js
@@ -1,13 +1,33 @@
-import generateLlmsPlugin from '../../../../../tdesign-uniapp/site/web/plugins/generate-llms';
+import path from 'path';
+
+import generateLlmsDocs from '../../../../../common/docs/plugins/generate-llms';
 
 /**
- * uniapp chat 站点专用：生成组件的 LLM Markdown 文档。
- * 组件目录为 uniapp-pro-components/chat。
+ * vite 插件：chat 站点构建时，基于 CHAT_COMPONENT_MAP 生成组件的 LLM Markdown 文档。
+ * 核心逻辑为纯 JS 方法 generateLlmsDocs，此处仅负责 vite 构建钩子分发。
  */
-export default function generateUniappChatLlmsPlugin() {
-  return generateLlmsPlugin({
-    componentsDir: '../../uniapp-pro-components/chat',
-    siteTitle: 'TDesign UniApp Chat',
-    siteDescription: 'TDesign UniApp Chat 组件库的 LLM 友好文档索引。',
-  });
+export default function generateChatLlmsPlugin() {
+  let config;
+  return {
+    name: 'generate-llms',
+    configResolved(resolvedConfig) {
+      config = resolvedConfig;
+    },
+    async closeBundle(error) {
+      if (error) return;
+      if (!config.env.PROD && config.env.MODE !== 'preview') return;
+
+      const siteRoot = config.root;
+      const componentsRoot = path.resolve(siteRoot, '../../pro-components/chat');
+      const outputDir = config.build.outDir || path.join(siteRoot, 'dist');
+
+      await generateLlmsDocs({
+        componentsRoot,
+        outputDir,
+        platform: 'chat',
+        siteTitle: 'TDesign Uniapp Chat',
+        siteDescription: 'TDesign Uniapp AI Chat 组件库的 LLM 友好文档索引。',
+      });
+    },
+  };
 }
