+++ packages/tdesign-uniapp/site/web/plugins/generate-llms/index.js
@@ -1,30 +1,12 @@
-import { readFileSync } from 'fs';
 import path from 'path';
 
-import generateLlmsDocs from '../../../../../common-docs/plugins/generate-llms/index.ts';
+import generateLlmsDocs from '../../../../../common/docs/plugins/generate-llms';
 
 /**
- * uniapp 站点专用：读取 demo 目录下的 index.vue 源码，输出 Vue SFC 代码块。
+ * vite 插件：chat 站点构建时，基于 CHAT_COMPONENT_MAP 生成组件的 LLM Markdown 文档。
+ * 核心逻辑为纯 JS 方法 generateLlmsDocs，此处仅负责 vite 构建钩子分发。
  */
-function readVueDemoCode(componentDir, demoName) {
-  const demoDir = path.join(componentDir, '_example', demoName);
-  const content = readFileSync(path.join(demoDir, 'index.vue'), 'utf-8');
-  if (!content.trim()) return '';
-  return ['```vue', content, '```'].join('\n');
-}
-
-/**
- * vite 插件：uniapp 站点构建时，为每个组件生成 LLM Markdown 文档。
- * 核心逻辑为纯 JS 方法 generateLlmsDocs。
- * 通过 options 覆盖 componentsDir / siteTitle / siteDescription / readDemoCode 可供 uniapp-chat 等站点复用。
- */
-export default function generateUniappLlmsPlugin(options = {}) {
-  const {
-    componentMap = {},
-    componentsDir = '../../uniapp-components',
-    siteTitle = 'TDesign UniApp',
-    siteDescription = 'TDesign UniApp 端组件库的 LLM 友好文档索引。',
-  } = options;
+export default function generateChatLlmsPlugin() {
   let config;
   return {
     name: 'generate-llms',
@@ -36,16 +18,15 @@ export default function generateUniappLlmsPlugin(options = {}) {
       if (!config.env.PROD && config.env.MODE !== 'preview') return;
 
       const siteRoot = config.root;
-      const componentsRoot = path.resolve(siteRoot, componentsDir);
+      const componentsRoot = path.resolve(siteRoot, '../../pro-components/chat');
       const outputDir = config.build.outDir || path.join(siteRoot, 'dist');
 
       await generateLlmsDocs({
         componentsRoot,
         outputDir,
-        componentMap,
-        siteTitle,
-        siteDescription,
-        readDemoCode: readVueDemoCode,
+        platform: 'chat',
+        siteTitle: 'TDesign Uniapp',
+        siteDescription: 'TDesign Uniapp 组件库的 LLM 友好文档索引。',
       });
     },
   };
